+++ app/models/analysis.rb
@@ -46,6 +46,10 @@ class Analysis < ApplicationRecord
       examples.map{|example| "#{example.title} #{example.description}"}
     end
 
+    text :attachments do 
+      attachments.map{|a| a.description}
+    end
+
     ## For scoping and faceting.
     integer :creator_facet do 
       creator_id