+++ config/routes.rb
@@ -56,6 +56,8 @@ Rails.application.routes.draw do
     # get    "#{resource_base}/:id/edit" => "attachments#edit"
     post   "#{resource_base}"          => "attachments#create"
     delete "#{resource_base}/:id"      => "attachments#destroy"
+    post "#{resource_base}/:id"        => "attachments#update"
+    put "#{resource_base}/reorder"    => "attachments#reorder"
   end
 
   ## Configures all of the routes for connecting two verticals together.